metal: keep compressor pair grid indices unsigned
Context:
- The paired E4M3 compressor kernel compiled and passed exactness on the M2 Ultra, but Metal emitted a signed/unsigned comparison warning for the z-grid selector.

Changes:
- Represent the compile-time batch count as uint and cast the final modulo result to int only after selection.

Rationale:
- Metal grid coordinates are unsigned. Matching their type removes the warning without changing generated grid selection or arithmetic.

Validation:
- The preceding candidate build compiled embedded Metal and the paired path passed 6/6 MTL0 exactness cases. A clean incremental embedded-Metal rebuild follows this commit.

Performance:
- Not performance-sensitive; type cleanup only.

Risks and tradeoffs:
- None identified. FC batch dimensions are positive int16 constants and their product fits uint and int for every supported matrix shape.

Follow-up:
- Rebuild embedded Metal and retain only if the isolated prompt-pair gain survives the final decision review.

Agent: Codex
Model: openai-codex/gpt-5.6-sol

+++ ggml/src/ggml-metal/ggml-metal.metal
@@ -12545,9 +12545,9 @@ kernel void kernel_mul_mm_e4m3_m2(
     threadgroup float  * lutg = (threadgroup float  *)(shmem + 4096 + NT1*2048);       // 256-entry f32 LUT (1024 B, style 0)
     threadgroup ushort * lutu = (threadgroup ushort *)(shmem + 4096 + NT1*2048);       // 256-entry bf16-bits LUT (512 B, style 2)
 
-    const int n_batch = FC_mul_mm_ne12*FC_mul_mm_ne13;
+    const uint n_batch = uint(FC_mul_mm_ne12)*uint(FC_mul_mm_ne13);
     const bool pair = PAIRED && tgpig.z >= n_batch;
-    const int im = PAIRED ? tgpig.z % n_batch : tgpig.z;
+    const int im = PAIRED ? int(tgpig.z % n_batch) : int(tgpig.z);
     device const char * src0_selected = pair ? src0_pair : src0;
     device       char * dst_selected  = pair ? dst_pair  : dst;
     const int r0 = tgpig.y*NR0;
